Avoid the divide-by-zero errors caused by non-short-circuit `if`
Newer version of GNU compiler more tightly conform to the Fortran standard
and do not immediately short circuit IF test processing once a statement
is known to be false.

This behavior causes divide-by-zero errors in module_ldas_noahmp.F90
- mod(itime, restart_frequency_hours * ...)
  if restart_frequency_hours == 0 for non-restart runs.
- mod(itime, stepwtd) if stepwtd == 0.

The two if tests in module_ldas_noahmp.F90 are re-arranged.
Now each of them are separated into two tests,
and the second test can be protected by the first one.