No longer ignore *.f90
Filenames in Mac OS X and Windows are case-insensitive.
The rule of *.f90 also matches *.F90 on such systems.

The make does not generate temperary files now.
The object files are directly compiled from the *.F90 source code now.

Thus, it is safe to remove the excluding rule of *.f90.